Create an abstract item schema
